+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Full changelog for Zend Framework 3 view helpers library.
 
+## v0.63.0 - Tests
+
+* Removing call to default escape plugins, assumption now is developer will escape their own data.
+* Added tests for Bootstrap 4 Badge view helper
+
 ## v0.62.1 - Beta 4.0.0 updates/fixes
 
 * Badge view helper updated, classes match beta-4.0.0 of Bootstrap.